==> notebook/1.table-cleanup.ipynb <==
# Crteate Recommender Schema and Tables

import pandas as pd

## Movies Metadata Table Preperation

movies_metadata = pd.read_csv('./data/movies_metadata.csv', low_memory=False)

movies_metadata[['id', 'title', 'release_date']]

movies_metadata = movies_metadata.iloc[:10000]

movies_metadata.shape

movies_metadata[['id', 'title', 'release_date']]

movies_metadata.dropna(subset = ['release_date'], inplace=True)

movies_metadata[movies_metadata['release_date'].isnull()]

movies_metadata.to_csv('./data/clean_movies_metadata.csv', index=False)

## Credits Table Preperation

credits_df = pd.read_csv('./data/credits.csv')

credits_df[credits_df['id'].isnull()]

credits_df.to_csv('./data/clean_credits.csv', index=False)

## Keywords Table Preperation

keywords_df = pd.read_csv('./data/keywords.csv')

keywords_df[keywords_df['id'].isnull()]

keywords_df.to_csv('./data/clean_keywords.csv', index=False)
