==> notebook/2.init-sql.ipynb <==
# Crteate Recommender Schema and Tables

## 1.Connecting to the database

%run 0.db_connection.ipynb

## 2. Movies Metadata Table Preperation

%%sql

DROP TABLE IF EXISTS movies_metadata CASCADE;

%%sql

DROP SCHEMA IF EXISTS recommender CASCADE;

%%sql

CREATE SCHEMA recommender;

%%sql

CREATE TABLE recommender.movies_metadata(
    adult TEXT,
    belongs_to_collection TEXT,
    budget TEXT,
    genres TEXT,
    homepage TEXT,
    id TEXT,
    imdb_id TEXT,
    original_language TEXT,
    original_title TEXT,
    overview TEXT,
    popularity TEXT,
    poster_path TEXT,
    production_companies TEXT,
    production_countries TEXT,
    release_date TEXT,
    revenue TEXT,
    runtime TEXT,
    spoken_languages TEXT,
    status TEXT,
    tagline TEXT,
    title TEXT,
    video TEXT,
    vote_average TEXT,
    vote_count TEXT
)

%%capture capture_output

%%sql

SET search_path TO recommender, public;

%%sql

SHOW search_path;

%%sql

\copy movies_metadata(adult,belongs_to_collection,budget,genres,homepage,id,imdb_id,original_language,original_title,overview,popularity,poster_path,production_companies,production_countries,release_date,revenue,runtime,spoken_languages,status,tagline,title,video,vote_average,vote_count)
FROM './data/clean_movies_metadata.csv'
DELIMITER ','
CSV HEADER;

%%sql

SELECT * FROM movies_metadata LIMIT 5;

%%sql

DELETE FROM movies_metadata WHERE id = '1997-08-20';
DELETE FROM movies_metadata WHERE id = '2012-09-29';

%%sql

SELECT id, COUNT(id), title
FROM movies_metadata
GROUP BY id, title
HAVING COUNT(id) > 1
ORDER BY COUNT(id);

%%sql

DELETE FROM movies_metadata
WHERE id IN
    (SELECT id
    FROM 
        (SELECT id,
         ROW_NUMBER() OVER( PARTITION BY id
        ORDER BY id ) AS row_num
        FROM movies_metadata ) t
        WHERE t.row_num > 1 );

%%sql

ALTER TABLE movies_metadata ALTER COLUMN id TYPE NUMERIC USING id::numeric;

%%sql

ALTER TABLE movies_metadata ADD CONSTRAINT movies_metadata_pk PRIMARY KEY (id);

%%sql

ALTER TABLE movies_metadata ALTER COLUMN adult TYPE BOOLEAN USING adult::boolean;

%%sql

ALTER TABLE movies_metadata ALTER COLUMN budget TYPE NUMERIC USING budget::numeric;

%%sql

ALTER TABLE movies_metadata ALTER COLUMN original_title TYPE VARCHAR(255);

%%sql

ALTER TABLE movies_metadata ALTER COLUMN popularity TYPE DECIMAL USING popularity::decimal;

%%sql

ALTER TABLE movies_metadata ALTER COLUMN release_date TYPE DATE USING release_date::date;

%%sql

ALTER TABLE movies_metadata ALTER COLUMN revenue TYPE MONEY USING revenue::money;

%%sql

ALTER TABLE movies_metadata ALTER COLUMN runtime TYPE DECIMAL USING runtime::decimal;

%%sql

ALTER TABLE movies_metadata ALTER COLUMN title TYPE VARCHAR(255);

%%sql

ALTER TABLE movies_metadata ALTER COLUMN vote_average TYPE DECIMAL USING vote_average::decimal;

%%sql

ALTER TABLE movies_metadata ALTER COLUMN vote_count TYPE DECIMAL USING vote_count::decimal;

## 3. Credits Table Preperation

%%sql

DROP TABLE IF EXISTS recommender.credits CASCADE;

%%sql

CREATE TABLE recommender.credits(
    "cast" TEXT,
    crew TEXT,
    id TEXT
)

%%sql

\copy recommender.credits("cast",crew,id)
FROM './data/clean_credits.csv'
DELIMITER ','
CSV HEADER;

%%sql

SELECT id, COUNT(id)
FROM recommender.credits
GROUP BY id
HAVING COUNT(id) > 1
ORDER BY COUNT(id);

%%sql

DELETE FROM recommender.credits
WHERE id IN
    (SELECT id
    FROM 
        (SELECT id,
         ROW_NUMBER() OVER( PARTITION BY id
        ORDER BY id ) AS row_num
        FROM recommender.credits ) t
        WHERE t.row_num > 1 );

%%sql

ALTER TABLE recommender.credits ADD CONSTRAINT credits_pk PRIMARY KEY (id);

%%sql

ALTER TABLE recommender.credits ALTER COLUMN id TYPE NUMERIC USING id::numeric;

%%sql

DELETE FROM recommender.credits 
    WHERE id NOT IN (
        SELECT ID FROM recommender.movies_metadata
    )

%%sql

ALTER TABLE recommender.credits
ADD CONSTRAINT id_fk FOREIGN KEY (id) REFERENCES recommender.movies_metadata(id);

## 4. Keywords Table Preperation

%%sql

DROP TABLE IF EXISTS recommender.keywords CASCADE;

%%sql

CREATE TABLE recommender.keywords(
    id TEXT,
    keywords TEXT
)

%%sql

\copy recommender.keywords(id,keywords)
FROM './data/clean_keywords.csv'
DELIMITER ','
CSV HEADER;

%%sql

SELECT id, COUNT(id)
FROM recommender.keywords
GROUP BY id
HAVING COUNT(id) > 1
ORDER BY COUNT(id);

%%sql

DELETE FROM recommender.keywords
WHERE id IN
    (SELECT id
    FROM 
        (SELECT id,
         ROW_NUMBER() OVER( PARTITION BY id
        ORDER BY id ) AS row_num
        FROM recommender.keywords ) t
        WHERE t.row_num > 1 );

%%sql

ALTER TABLE recommender.keywords ADD CONSTRAINT keywords_pk PRIMARY KEY (id);

%%sql

ALTER TABLE recommender.keywords ALTER COLUMN id TYPE NUMERIC USING id::numeric;

%%sql

DELETE FROM recommender.keywords 
    WHERE id NOT IN (
        SELECT ID FROM recommender.movies_metadata
    )

%%sql

ALTER TABLE recommender.keywords
ADD CONSTRAINT id_fk FOREIGN KEY (id) REFERENCES recommender.movies_metadata(id);
