==> notebook/3.metadata-based-recommender.ipynb <==
# Importing libraries
import pandas as pd
from ast import literal_eval
import numpy as np
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.metrics.pairwise import cosine_similarity
import pickle
import blosc

### Connecting to the database

%run 0.db_connection.ipynb

### Prepere Movies Matadata

%%sql movies_metadata <<

SELECT * FROM recommender.movies_metadata;

movies_metadata.sample()

movies_metadata.columns

movies_metadata.shape

movies_metadata.dtypes

### Prepere Keywords

%%sql keywords <<

SELECT * FROM recommender.keywords;

keywords.dtypes

keywords.sample()

### Prepere Credits

%%sql credits <<

SELECT * FROM recommender.credits;

credits.dtypes

credits.sample()

### Make soup out of metadata

df = movies_metadata.merge(credits, on='id')

df = df.merge(keywords, on='id')

df.columns

features = ['cast', 'crew', 'keywords', 'genres']
for feature in features:
    df[feature] = df[feature].apply(literal_eval)

df.sample()

df.iloc[6567]['crew'][0]

type(df.iloc[6567]['crew'][0])

def get_director(x):
    for crew_member in x:
        if crew_member['job'] == 'Director':
            return crew_member['name']
    return np.nan

df['director'] = df['crew'].apply(get_director)

def generate_list(x):
    if isinstance(x, list):
        names = [i['name'] for i in x]
        return names
        
    return []

df['cast'] = df['cast'].apply(generate_list)
df['keywords'] = df['keywords'].apply(generate_list)
df['genres'] = df['genres'].apply(generate_list)

df[['title', 'cast', 'director', 'keywords', 'genres']].sample()

# Function to sanitize data to prevent ambiguity. It removes spaces and converts to lowercase
def sanitize(x):
    if isinstance(x, list):
        #Strip spaces and convert to lowercase
        return [str.lower(i.replace(" ", "")) for i in x]
    else:
        #Check if director exists. If not, return empty string
        if isinstance(x, str):
            return str.lower(x.replace(" ", ""))
        else:
            return ''

for feature in ['cast', 'director', 'genres', 'keywords']:
    df[feature] = df[feature].apply(sanitize)

df.sort_values(by=['release_date'], inplace=True, ascending=False, na_position='first')

df[['id', 'title', 'release_date']]

df[['id', 'title', 'cast', 'director', 'keywords', 'genres']].sample()

#Function that creates a soup out of the desired metadata
def create_soup(x):
    return ' '.join(x['keywords']) + ' ' + ' '.join(x['cast']) + ' ' + x['director'] + ' ' + ' '.join(x['genres'])

# Create the new soup feature
df['soup'] = df.apply(create_soup, axis=1)

df.iloc[5842]['soup']

### Prepere the model

count = CountVectorizer(stop_words='english')

count_matrix = count.fit_transform(df['soup'])

count_matrix = count_matrix.astype(np.float32)

cosine_sim = cosine_similarity(count_matrix, count_matrix)

df = df.reset_index()

indices = pd.Series(df.index, index=df['title'])

def content_recommender(title, cosine_sim=cosine_sim, df=df, indices=indices):
    # Obtain the index of the movie that matches the title
    idx = indices[title]

    # Get the pairwsie similarity scores of all movies with that movie
    # And convert it into a list of tuples as described above
    sim_scores = list(enumerate(cosine_sim[idx]))

    # Sort the movies based on the cosine similarity scores
    sim_scores = sorted(sim_scores, key=lambda x: x[1], reverse=True)

    # Get the scores of the 10 most similar movies. Ignore the first movie.
    sim_scores = sim_scores[1:11]

    # Get the movie indices
    movie_indices = [i[0] for i in sim_scores]

    # Return the top 10 most similar movies
    return df['title'].iloc[movie_indices]

content_recommender('The Lion King')

content_recommender('Star Wars')

### Generate compresses pickle file

pickle_path_pkl = 'mlmodels/cosine_similarity_model.pkl'

with open(pickle_path_pkl, 'wb') as pickle_file:
    pickle.dump(cosine_sim, pickle_file, protocol=4)

pickle_path = 'mlmodels/cosine_similarity_model.dat'

pickled_data = pickle.dumps(cosine_sim, protocol=4)

compressed_pickle = blosc.compress(pickled_data)

with open(pickle_path, 'wb') as file:
    file.write(compressed_pickle)
